refactor(components/streakCalendar): turn streak calendar into a component
