+++ Front-end/src/components/index.tsx
@@ -3,4 +3,5 @@ export { default as AppHeader } from './AppHeader';
 export { default as Button } from './Button';
 export { default as Footer } from './Footer';
 export { default as H1 } from './H1';
+export { default as StreakCalendar } from './StreakCalendar';
 export { default as TextBox } from './TextBox';
